Make photo detail list be in a bottom sheet
Had to pull in the BottomSheet library, because the native `.sheet`
makes the sheet cover the bottom bar. This is also nicer though since
the built in sheet's `.medium` size might've covered the user icon on
the map. Also I don't remember if it was translucent, but translucency
is cool, so I'm all for that.