UI and UX work (#1)
* Make photo detail list be in a bottom sheet

Had to pull in the BottomSheet library, because the native `.sheet`
makes the sheet cover the bottom bar. This is also nicer though since
the built in sheet's `.medium` size might've covered the user icon on
the map. Also I don't remember if it was translucent, but translucency
is cool, so I'm all for that.

* Break apart index view into smaller bits

* Get rid of spacing between map and details

* Adjust map height

* Ability to copy values

* Sync packages

IDK why this changed. Fucking Xcode.