Remove ViewInspector tests
ViewInspector seems to fall apart when I'm trying to test a more complex
view with interaction (i.e. `IndexViews`). I'd like it to work, but I
don't want to keep banging my head on it. I also prefer the API of UI
tests, and think I'll just suck up the slowness and fact it has to be a
*complete* black box.